requirements.txt
pydantic
pydantic-ai
pypdf
requests
python-dotenv
pandas
llmwhisperer-client